Musterpoint-GCL Hub IKEJA/Musterpoint Ogba/Tolulope Odueke [Lagos Trip Curator - Python Functionalities]/requirements.txt
bs4==0.0.1
weather-api==1.0.4